+++ sql/core/src/main/scala/org/apache/spark/sql/execution/command/commands.scala
@@ -17,11 +17,8 @@
 
 package org.apache.spark.sql.execution.command
 
-import java.util.NoSuchElementException
-
-import org.apache.spark.internal.Logging
 import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.{AnalysisException, Dataset, Row, SQLContext}
+import org.apache.spark.sql.{Dataset, Row, SQLContext}
 import org.apache.spark.sql.catalyst.{CatalystTypeConverters, InternalRow, TableIdentifier}
 import org.apache.spark.sql.catalyst.errors.TreeNodeException
 import org.apache.spark.sql.catalyst.expressions.{Attribute, AttributeReference}
@@ -29,7 +26,6 @@ import org.apache.spark.sql.catalyst.plans.logical
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.execution.SparkPlan
 import org.apache.spark.sql.execution.debug._
-import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.types._
 
 /**
@@ -76,156 +72,6 @@ private[sql] case class ExecutedCommand(cmd: RunnableCommand) extends SparkPlan 
   override def argString: String = cmd.toString
 }
 
-
-case class SetCommand(kv: Option[(String, Option[String])]) extends RunnableCommand with Logging {
-
-  private def keyValueOutput: Seq[Attribute] = {
-    val schema = StructType(
-      StructField("key", StringType, false) ::
-        StructField("value", StringType, false) :: Nil)
-    schema.toAttributes
-  }
-
-  private val (_output, runFunc): (Seq[Attribute], SQLContext => Seq[Row]) = kv match {
-    // Configures the deprecated "mapred.reduce.tasks" property.
-    case Some((SQLConf.Deprecated.MAPRED_REDUCE_TASKS, Some(value))) =>
-      val runFunc = (sqlContext: SQLContext) => {
-        logWarning(
-          s"Property ${SQLConf.Deprecated.MAPRED_REDUCE_TASKS} is deprecated, " +
-            s"automatically converted to ${SQLConf.SHUFFLE_PARTITIONS.key} instead.")
-        if (value.toInt < 1) {
-          val msg =
-            s"Setting negative ${SQLConf.Deprecated.MAPRED_REDUCE_TASKS} for automatically " +
-              "determining the number of reducers is not supported."
-          throw new IllegalArgumentException(msg)
-        } else {
-          sqlContext.setConf(SQLConf.SHUFFLE_PARTITIONS.key, value)
-          Seq(Row(SQLConf.SHUFFLE_PARTITIONS.key, value))
-        }
-      }
-      (keyValueOutput, runFunc)
-
-    case Some((SQLConf.Deprecated.EXTERNAL_SORT, Some(value))) =>
-      val runFunc = (sqlContext: SQLContext) => {
-        logWarning(
-          s"Property ${SQLConf.Deprecated.EXTERNAL_SORT} is deprecated and will be ignored. " +
-            s"External sort will continue to be used.")
-        Seq(Row(SQLConf.Deprecated.EXTERNAL_SORT, "true"))
-      }
-      (keyValueOutput, runFunc)
-
-    case Some((SQLConf.Deprecated.USE_SQL_AGGREGATE2, Some(value))) =>
-      val runFunc = (sqlContext: SQLContext) => {
-        logWarning(
-          s"Property ${SQLConf.Deprecated.USE_SQL_AGGREGATE2} is deprecated and " +
-            s"will be ignored. ${SQLConf.Deprecated.USE_SQL_AGGREGATE2} will " +
-            s"continue to be true.")
-        Seq(Row(SQLConf.Deprecated.USE_SQL_AGGREGATE2, "true"))
-      }
-      (keyValueOutput, runFunc)
-
-    case Some((SQLConf.Deprecated.TUNGSTEN_ENABLED, Some(value))) =>
-      val runFunc = (sqlContext: SQLContext) => {
-        logWarning(
-          s"Property ${SQLConf.Deprecated.TUNGSTEN_ENABLED} is deprecated and " +
-            s"will be ignored. Tungsten will continue to be used.")
-        Seq(Row(SQLConf.Deprecated.TUNGSTEN_ENABLED, "true"))
-      }
-      (keyValueOutput, runFunc)
-
-    case Some((SQLConf.Deprecated.CODEGEN_ENABLED, Some(value))) =>
-      val runFunc = (sqlContext: SQLContext) => {
-        logWarning(
-          s"Property ${SQLConf.Deprecated.CODEGEN_ENABLED} is deprecated and " +
-            s"will be ignored. Codegen will continue to be used.")
-        Seq(Row(SQLConf.Deprecated.CODEGEN_ENABLED, "true"))
-      }
-      (keyValueOutput, runFunc)
-
-    case Some((SQLConf.Deprecated.UNSAFE_ENABLED, Some(value))) =>
-      val runFunc = (sqlContext: SQLContext) => {
-        logWarning(
-          s"Property ${SQLConf.Deprecated.UNSAFE_ENABLED} is deprecated and " +
-            s"will be ignored. Unsafe mode will continue to be used.")
-        Seq(Row(SQLConf.Deprecated.UNSAFE_ENABLED, "true"))
-      }
-      (keyValueOutput, runFunc)
-
-    case Some((SQLConf.Deprecated.SORTMERGE_JOIN, Some(value))) =>
-      val runFunc = (sqlContext: SQLContext) => {
-        logWarning(
-          s"Property ${SQLConf.Deprecated.SORTMERGE_JOIN} is deprecated and " +
-            s"will be ignored. Sort merge join will continue to be used.")
-        Seq(Row(SQLConf.Deprecated.SORTMERGE_JOIN, "true"))
-      }
-      (keyValueOutput, runFunc)
-
-    case Some((SQLConf.Deprecated.PARQUET_UNSAFE_ROW_RECORD_READER_ENABLED, Some(value))) =>
-      val runFunc = (sqlContext: SQLContext) => {
-        logWarning(
-          s"Property ${SQLConf.Deprecated.PARQUET_UNSAFE_ROW_RECORD_READER_ENABLED} is " +
-            s"deprecated and will be ignored. Vectorized parquet reader will be used instead.")
-        Seq(Row(SQLConf.PARQUET_VECTORIZED_READER_ENABLED, "true"))
-      }
-      (keyValueOutput, runFunc)
-
-    // Configures a single property.
-    case Some((key, Some(value))) =>
-      val runFunc = (sqlContext: SQLContext) => {
-        sqlContext.setConf(key, value)
-        Seq(Row(key, value))
-      }
-      (keyValueOutput, runFunc)
-
-    // (In Hive, "SET" returns all changed properties while "SET -v" returns all properties.)
-    // Queries all key-value pairs that are set in the SQLConf of the sqlContext.
-    case None =>
-      val runFunc = (sqlContext: SQLContext) => {
-        sqlContext.getAllConfs.map { case (k, v) => Row(k, v) }.toSeq
-      }
-      (keyValueOutput, runFunc)
-
-    // Queries all properties along with their default values and docs that are defined in the
-    // SQLConf of the sqlContext.
-    case Some(("-v", None)) =>
-      val runFunc = (sqlContext: SQLContext) => {
-        sqlContext.conf.getAllDefinedConfs.map { case (key, defaultValue, doc) =>
-          Row(key, defaultValue, doc)
-        }
-      }
-      val schema = StructType(
-        StructField("key", StringType, false) ::
-          StructField("default", StringType, false) ::
-          StructField("meaning", StringType, false) :: Nil)
-      (schema.toAttributes, runFunc)
-
-    // Queries the deprecated "mapred.reduce.tasks" property.
-    case Some((SQLConf.Deprecated.MAPRED_REDUCE_TASKS, None)) =>
-      val runFunc = (sqlContext: SQLContext) => {
-        logWarning(
-          s"Property ${SQLConf.Deprecated.MAPRED_REDUCE_TASKS} is deprecated, " +
-            s"showing ${SQLConf.SHUFFLE_PARTITIONS.key} instead.")
-        Seq(Row(SQLConf.SHUFFLE_PARTITIONS.key, sqlContext.conf.numShufflePartitions.toString))
-      }
-      (keyValueOutput, runFunc)
-
-    // Queries a single property.
-    case Some((key, None)) =>
-      val runFunc = (sqlContext: SQLContext) => {
-        val value =
-          try sqlContext.getConf(key) catch {
-            case _: NoSuchElementException => "<undefined>"
-          }
-        Seq(Row(key, value))
-      }
-      (keyValueOutput, runFunc)
-  }
-
-  override val output: Seq[Attribute] = _output
-
-  override def run(sqlContext: SQLContext): Seq[Row] = runFunc(sqlContext)
-
-}
 
 /**
  * An explain command for users to see how a command will be executed.
@@ -307,22 +153,6 @@ case object ClearCacheCommand extends RunnableCommand {
   override def output: Seq[Attribute] = Seq.empty
 }
 
-
-case class DescribeCommand(
-    table: TableIdentifier,
-    override val output: Seq[Attribute],
-    isExtended: Boolean)
-  extends RunnableCommand {
-
-  override def run(sqlContext: SQLContext): Seq[Row] = {
-    val relation = sqlContext.sessionState.catalog.lookupRelation(table)
-    relation.schema.fields.map { field =>
-      val cmtKey = "comment"
-      val comment = if (field.metadata.contains(cmtKey)) field.metadata.getString(cmtKey) else ""
-      Row(field.name, field.dataType.simpleString, comment)
-    }
-  }
-}
 
 /**
  * A command for users to get tables in the given database.
